introduces the notions of Measurement and MeasurementDelta
a `Measurement` is a a given timestamp (which can measure multiple things)
a `MeasurementDelta` is the difference between two `Measurements` and
currently trackes real-time and the ratio of cpu-time/real-time.

All the conversion and calculation is part of MeasurementDelta so that it is 'cheap'
to obtain a Measurement. As an example currently we convert every `time_ns` to Float64
which we can delay using this notion of a MeasurementDelta.