Improve/fix show methods
Previously the logic for switching compact or verbose form of various
benchmark result objects are in `show(io::IO, group::BenchmarkGroup)`.
A more natural place would be `show(io::IO, t::Trial)` etc. as it let
code outside BenchmarkTools to switch the compactness using the
standard `IOContext` interface.  For example, with this patch `Trial`s
in `Vector` are shown in the compact form automatically.

This patch also adds:
* Colored output for the words `improvement` and `regression`.
* `:typeinfo` IO context support