Fix how IOContext :limit is handled
The docstring of `IOContext` specifies that `:limit` contains a `Bool`
for toggling if the output should be truncated.  The previous usage of
`:limit` in BenchmarkTools.jl treats it as an upper bound of the
numebr of elements to be shown; it is incompatible with the common
usage.  This PR fixes it by introducing a new (non-standard) context
`:boundto` that can be used to specify the upper bound of the number
of elements to be printed for each level.  The old usage of `:limit`
is still supported with deprecation if it is not a `Bool`.