Add `@bprofile` (#254)
To collect enough data, profiling benefits from long(ish) execution
times. That makes `@benchmark`/`@btime` attractive workloads
for profiling. Unfortunately, there's a lot of "junk" that appears in
the profile records, including many for type-inference.

This adds `@bprofile`, a substitute for `@profile @b...`,
that restricts profiling to the `run` phase of benchmarking.
Other than a fairly small amount of inference on the first session
use of `@bprofile` itself, this eliminates all the inference
and tuning in the trace.